Added all endpoints to create/update/remove for orders and carts

+++ src/main/java/com/harriet/shopiify/cart/dto/CartItemDTO.java
@@ -20,7 +20,7 @@ public class CartItemDTO {
 
     private String units;
 
-    private BigInteger stock;
+    private Long stock;
 
     private Double price;
